+++ openchain/consensus/pbft/viewchange.go
@@ -171,7 +171,7 @@ func (instance *Plugin) recvViewChange(vc *ViewChange) error {
 			minView = idx.v
 		}
 	}
-	if uint(len(replicas)) >= instance.f+1 {
+	if len(replicas) >= instance.f+1 {
 		logger.Info("Replica %d received f+1 view-change messages, triggering view-change to view %d",
 			instance.id, minView)
 		// subtract one, because sendViewChange() increments
@@ -363,13 +363,13 @@ func (instance *Plugin) selectInitialCheckpoint(vset []*ViewChange) (checkpoint 
 
 	for idx, vcList := range checkpoints {
 		// need weak certificate for the checkpoint
-		if uint(len(vcList)) <= instance.f { // type casting necessary to match types
+		if len(vcList) <= instance.f { // type casting necessary to match types
 			logger.Debug("no weak certificate for n:%d",
 				idx.SequenceNumber)
 			continue
 		}
 
-		quorum := uint(0)
+		quorum := 0
 		for _, vc := range vcList {
 			if vc.H <= idx.SequenceNumber {
 				quorum++
@@ -403,7 +403,7 @@ nLoop:
 		for _, m := range vset {
 			// "...with <n,d,v> ∈ m.P"
 			for _, em := range m.Pset {
-				quorum := uint(0)
+				quorum := 0
 				// "A1. ∃2f+1 messages m' ∈ S"
 			mpLoop:
 				for _, mp := range vset {
@@ -446,7 +446,7 @@ nLoop:
 			}
 		}
 
-		quorum := uint(0)
+		quorum := 0
 		// "else if ∃2f+1 messages m ∈ S"
 	nullLoop:
 		for _, m := range vset {
